requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
bidict==0.23.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
cloudinary==1.44.1
Django==6.0.3
django-cloudinary-storage==0.3.0
django-environ==0.13.0
django-solo==2.5.1
docutils==0.22.4
fastapi==0.125.0
frozenlist==1.8.0
gunicorn==25.1.0
h11==0.16.0
htmx==0.0.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown2==2.5.5
MarkupSafe==3.0.3
multidict==6.7.1
nicegui==3.8.0
orjson==3.11.7
packaging==26.0
pillow==12.1.1
propcache==0.4.1
pydantic==1.10.26
pydantic_core==2.42.0
Pygments==2.19.2
python-dotenv==1.2.2
python-engineio==4.13.1
python-magic==0.4.27
python-multipart==0.0.22
python-socketio==5.16.1
PyYAML==6.0.3
requests==2.32.5
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.5
starlette==0.50.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wsproto==1.3.2
yarl==1.23.0
